[refactor] 도메인 유효성 검증을 메서드로 분리하라
이미지, 강의 도메인 생성자의 유효성 검사를 메서드로 분리합니다
