+++ src/main/java/nextstep/courses/domain/course/image/Image.java
@@ -1,7 +1,6 @@
 package nextstep.courses.domain.course.image;
 
 import java.time.LocalDateTime;
-import java.util.Arrays;
 
 public class Image {
     public static final int MB = 1024 * 1024;
@@ -23,41 +22,6 @@ public class Image {
 
     private LocalDateTime updatedAt;
 
-    public enum Type {
-        GIF("gif"),
-        JPG("jpg"),
-        JPEG("jpeg"),
-        PNG("png"),
-        SVG("svg");
-
-        Type(String description) {
-            this.description = description;
-        }
-
-        private final String description;
-
-        public static Type find(String name) {
-            return Arrays.stream(values())
-                    .filter(imageType -> imageType.description.equals(name))
-                    .findAny()
-                    .orElseThrow(
-                            () -> new IllegalArgumentException(
-                                    String.format("허용하는 확장자는 다음과 같습니다.\n %s", descriptions())
-                            )
-                    );
-        }
-
-        public static String descriptions() {
-            StringBuilder sb = new StringBuilder();
-            for (Type type : values()) {
-                sb.append(type.description).append(", ");
-            }
-            sb.setLength(sb.length() - 2);
-
-            return sb.toString();
-        }
-    }
-
     public Image() {
     }
 
@@ -67,19 +31,9 @@ public class Image {
 
     public Image(Long id, int imageSize, String type, int imageWidth, int imageHeight,
                  LocalDateTime createdAt, LocalDateTime updatedAt) {
-        if (imageSize > 1 * MB) {
-            throw new IllegalArgumentException("사진 크기는 1MB를 넘을 수 없습니다.");
-        }
-
-        if(imageWidth < WIDTH_MIN || imageHeight < HEIGHT_MIN) {
-            throw new IllegalArgumentException(
-                    String.format("가로 픽셀은 %d, 세로 픽셀은 %d 이상이어야 합니다.", WIDTH_MIN, HEIGHT_MIN)
-            );
-        }
-
-        if((double) imageWidth / imageHeight != WIDTH_HEIGHT_RATIO) {
-            throw new IllegalArgumentException("가로 세로 비율은 3:2여야 합니다.");
-        }
+        checkImageSizeIsValid(imageSize);
+        checkWidthAndHeightSizeIsValid(imageWidth, imageHeight);
+        checkWidthAndHeightRatioIsValid(imageWidth, imageHeight);
 
         this.id = id;
         this.imageSize = imageSize;
@@ -88,5 +42,25 @@ public class Image {
         this.imageHeight = imageHeight;
         this.createdAt = createdAt;
         this.updatedAt = updatedAt;
+    }
+
+    private static void checkImageSizeIsValid(int imageSize) {
+        if (imageSize > 1 * MB) {
+            throw new IllegalArgumentException("사진 크기는 1MB를 넘을 수 없습니다.");
+        }
+    }
+
+    private static void checkWidthAndHeightSizeIsValid(int imageWidth, int imageHeight) {
+        if (imageWidth < WIDTH_MIN || imageHeight < HEIGHT_MIN) {
+            throw new IllegalArgumentException(
+                    String.format("가로 픽셀은 %d, 세로 픽셀은 %d 이상이어야 합니다.", WIDTH_MIN, HEIGHT_MIN)
+            );
+        }
+    }
+
+    private static void checkWidthAndHeightRatioIsValid(int imageWidth, int imageHeight) {
+        if ((double) imageWidth / imageHeight != WIDTH_HEIGHT_RATIO) {
+            throw new IllegalArgumentException("가로 세로 비율은 3:2여야 합니다.");
+        }
     }
 }
